Revert "Fixes incorrect install names on darwin platforms"
This reverts commit b06f49507e02384cb4ff1ff862da2a673e9a081f.
This commit changed the runfile_solib_name variable to match the
variable name that's expected at runtime in most cases.
However, in apollo we want to use the mangled path in bazel-out as an
SONAME in order to avoid calling install_name_tool in the clang wrapper
and invalidate the signature.

Issue: BEN-13563  Upgrade apollo to bazel 5
Change-Id: I5569143f36826f44446663a7623f535212fe5112
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/345653
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Paul (SPG) Janzen <paul_janzen@apple.com>