Remove '@' from RepositoryName constructors
This CL changes it so that RepositoryName#create no longer takes an argument that starts with '@' (that is, it now always takes the "stripped" name).

Also #createWithValidStrippedName is renamed to #createUnvalidated.

Notably, it does *not* try to change the caching logic, and still stores a '@'-prefixed string. This is in order to minimize change in behavior. Fixing those will come later.

PiperOrigin-RevId: 441759032