Add Blaze support for "local" optimization of individual Java libraries.
This should allow optimization on a per-library level within Blaze of every
artifact generated by javac. The purpose of this is for improved test coverage
by the AppReduce team and is not intended for general use. See this thread for
more context:

[]

This change should only add a dependency on the bytecode optimizer when
--experimental_local_java_optimizations is true, so it should be a no-op for
all standard builds.

I implemented this in Java instead of Starlark because it allowed me to modify
the produced JavaInfo much more simply. If this was implemented in Starlark,
I'd have to be able to replace the JavaInfo object produced by a java_* target
since the JavaInfo from javac is currently all still wired up in Java.  My
attempts at structuring the change that way seemed much more invasive to
JavaInfo and more brittle.

PiperOrigin-RevId: 441814811