Don't set BAZEL_USE_CPP_ONLY_TOOLCHAIN=1
This option was inherited from apollo when updating tools/bazel, but it
causes bazel to not detect the Xcode toolchain when building on MacOS.
This is desirable in apollo as we have our own logic, but not when
building bazel. We're reverting back to the default behavior here by
removing it.

Issue: BEN-13563  Upgrade apollo to bazel 5
Change-Id: I4bf8a5c4982060868a2f6239163da21005dc8fda
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/347176
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Paul (SPG) Janzen <paul_janzen@apple.com>