Restore old execution log diffing docs
While the new json execution log file is handy in some cases, it doesn't
provide the same diffing behavior as the binary file and the execlog
parser tool. The json file contains `walltime` attributes for each
action that (likely) differ per run, and it's not in any guaranteed
order.

This is a partial revert of b35944e4fedbfa65a7b081021959584ab9a07a84.

Closes #15274.

PiperOrigin-RevId: 445149528