ijar: read Code attribute
JVMS 4.7.3: "If the method is either native or abstract, its method_info
structure must not have a Code attribute. Otherwise, its method_info structure
must have exactly one Code attribute."

As hack around this, we instead emit a (dummy) Code attribute that does not
depend on the actual implementation of the method. This should fix compatibility
with tools like https://spotbugs.github.io which seem to care about the Code
attribute of dependencies and currently fail with ijar enabled.

Since there are some performance concerns about emitting a code attribute at
all, we start with only reading the Code attribute in this PR and add the code
to emit the dummy attributes in a follow-up to make benchmarking easier.

Progress on https://github.com/bazelbuild/bazel/issues/13546