Allow passing a minimal version to use for skyframe evaluations.
Allows the evaluator to specify a version compatible with its versioning scheme. This minimal version is used for the MTSV of hermetic nodes with no non-hermetic dependencies (e.g. `FileSymlinkCycleUniquenessFunction`) instead of the previously used `null`. This way it can be distinguished from a transient error which also uses `null`.

`InMemoryNodeEntry` no longer considers the MTSV, since no evaluator implementations depend on that behavior.

PiperOrigin-RevId: 445933706