Add tableOfContents to stardoc
This adds a new replacement in stardoc headers for a simple table of
contents that prints each different type of documented symbol, and a
link to each one that will be on the page. This is much easier than
having consumers enumerate all the symbols in a file to build this up
themselves.