Implement targetsToSkip for Skymeld.
To figure out which targets to skip, we have to use TopLevelConstraintSemantics, which is extra-Skyframe, and itself makes Skyframe evaluations. In order to make it work with Skymeld, we adopted the various checks in TopLevelConstraintSemantics into BuildDriverFunction.
While porting these checks to Skyframe, we tried as much as possible to reuse the code, hence the refactoring of TopLevelConstraintSemantics.

PiperOrigin-RevId: 446662371