Wire up CommandAdjuster in StarlarkCustomCommandLine
A `CommandAdjuster` is added to the `StarlarkSemantics` of
`Args#add_all` `map_each` callbacks and automatically applies path
mapping to the `path`, `dirname`, and `root` methods of `File`s.

In a follow-up commit, this can also be used to emit warnings if rules
hardcode the values of these struct fields during analysis, where their
mapped value is not yet available.