Delete Ninja support.
The plan was to use it with building Android (the operating system), but that project took another direction and thus parsing Ninja graphs will not be necessary anymore.

If this will ever be needed again, we can resurrect it from source control.

The whole functionality is behind the --experimental_ninja_actions experimental flag, so it can be deleted without warning.

RELNOTES: None.
PiperOrigin-RevId: 447397970