Do not necessarily watch for changes under '/'.
Before, we requested the root of external repositories using a Root of '/', which made Bazel climb up the directory tree until '/' when resolving symlinks when statting the locations where remote repositories are fetched. This is something that is not done for the main repository.

For the sake of efficiency and consistency, treat the actual location of external repositories the same as the main repository and do not register paths above their roots in Skyframe.

In addition, special-case the `@bazel_tools` repository so that the aforementioned ascent does not happen for its source (which should happen for every ther local repository).

RELNOTES: None.
PiperOrigin-RevId: 447409159