Make `coverage --combined_report=lcov` skip incompatible tests
Before this patch we see coverage trying to execute incompatible
tests. This results in `Can't build this. This target is incompatible`
messages. Instead, those tests should just be skipped.

This patch makes it so the coverage code only tries to collect
coverage artifacts for compatible tests. Artifacts for incompatible
tests are ignored.

Fixes #15385

Closes #15419.

PiperOrigin-RevId: 447710011