Allow any CcInfo providing rule in objc_library deps
Previously it was invalid to have any custom rules in the deps of
objc_library unless they provided `apple_common.Objc` which very few
rules do, but they do provide `CcInfo` for this type of integration. I
hit this issue with a `rust_library` being depended on by an
`objc_library`.