Remote: Fix the issue that partial downloaded inputs are not deleted if the request is cancelled.
Following https://github.com/bazelbuild/bazel/commit/280ef6915b0f507218a073974825d6aa7effddee, this change fixes the issue that partial downloaded files are not deleted. The root cause is, even we cancel the download futures inside AbstractActionInputPrefetcher and then delete the files, the actual downloads inside GrpcCacheClient are not cancelled so it is still writing to the path.

PiperOrigin-RevId: 447970542