Add `add_exports` and `add_opens` to `java_` rules
e.g.

```
java_library(
    ...
    add_exports = [
        "--add-exports=jdk.compiler/com.sun.tools.javac.api",
    ],
)
```

At compile-time the attributes correspond to javac's `--add-exports=` and
`--add-opens=` flags, to allow code in the unnamed module to access the given
packages.

At runtime the the attributes correspond to the similar runtime flags. They
are set in the stub script, and recorded as `Add-Exports:` and `Add-Opens:`
entries in the deploy jar's `MANIFEST.MF`.

PiperOrigin-RevId: 448039045