Implement `--experimental_content_based_path_mapping`
Starlark actions with mnemonics listed in the flag have their input
and output paths mapped according to the following scheme when staged
for sandboxed or remote execution:

* source file paths are preserved as is;
* paths of generated input files of the form
  `bazel-out/k8-fastbuild/bin/pkg/file` are replaced with
  `bazel-out/<hash>/bin/pkg/file`, where `<hash>` is a digest of the
  contents of the file;
* paths of output files of the form
  `bazel-out/k8-fastbuild/bin/pkg/file` are replaced with
  `bazel-out/out/bin/pkg/file`.

This path mapping is applied both to the location at which files are
staged in the sandbox or during remote execution and to paths emitted
into structured command lines. In order to build successfully with this
flag, action command lines

* must not contain `dirname`, `path`, or `root.path` of any `File`
  object evaluated at analysis time;
* must not reference `File` objects that aren't declared inputs or
  outputs of the action.

Spawns derived from actions opted into content-based path mapping can
only be executed in symlink-based sandboxes or remotely.