Initial import of rules_license.
This is a pure, third_party only initial import. A future one will hook into the Bazel WORKSPACE.

This is a pull from https://github.com/bazelbuild/rules_license/releases/tag/0.0.2,
with significant vendoring in to relatives the 'srcs' construct for Bazel.

Since this repository must become an external repository name
rules_license, it follows the model of googleapis and remoteapis.
Thus, the visibility rules and use of self referential names is
differnt from most other imports.

Part 1 of #15507