Revert test_setup.sh signal handling
Revert "Update test-setup.sh to in fact forward signals to its children.
It claims to in at least one mode, but actually implicitly relies on the
fact that the sandbox does. Second attempt: this accounts for allowing
test processes to receive SIGINT if they want to, which is harder than
it sounds."

This reverts commit 9051faa313f29b8ad8ded68b48b3f7af13108d77. See:
https://github.com/bazelbuild/bazel/commit/9051faa313f29b8ad8ded68b48b3f7af13108d77

Issue: OS-30019
Org-Issue: OS-27095
Change-Id: I0cdd2e72584a421a4965fe45259fecb10b234240
Reviewed-on: https://gerrit.sg.apple.com/c/swplatform/bazelbuild/bazel/+/352482
Tested-by: Hyperion Jenkins <hyperion_jenkins@apple.com>
Reviewed-by: Dave Inglis <dinglis@apple.com>