Allow skipping Xcode locator runs
For builds that require Xcode, running xcode_locator during toolchain
setup is potentially duplicative. The only things it's used for is to
check that they exist if it's expected, and to add to the allowed
include directories. The include directories already includes
`/Applications` which is where _most_ people likely have Xcode
installed. This saves a few seconds when the toolchain is reconfigured.