Explicitly mention that `alias()` ignores `testonly`.
This is to make clear that testonly is ignored to prevent accidental misuse
which is otherwise hard to detect. Exposing a testonly alias is an appealing
option when the non-testonly target is visibility-restricted, but usage
from arbitrary tests is acceptable (e.g., a fake/no-op implementation of
an authentication/authorization library could freely be used by any test,
but the underlying, non-testonly, library uses visibility so that only
vetted/trusted users can use it).

PiperOrigin-RevId: 449589636