Apply TopLevelStatusEvents to non-Skymeld code path.
So far we've only applied (TopLevelTarget|Aspect)BuiltEvent to the non-Skymeld
code path. This CL applies the rest of the events. Note that we post the events regardless of the value of --experimental_use_event_based_build_completion_status: it should only affect whether we use the result of BuildResultListener in the final result summary.

Also added integration tests for these.

PiperOrigin-RevId: 449685522