Bzlmod: Proper repo mappings for @_builtins
@_builtins is not a real repo and does not currently have repo mappings. However, things in there still sometimes use things from other repos (eg. @platforms). This isn't a problem pre-Bzlmod since @platforms is just called @platforms, but in Bzlmod it's called @platforms.0.0.4, so things break.

This CL changes it so that when Bzlmod is enabled, @_builtins adopts the repo mapping of @bazel_tools, the built-in module. Note that the repo mapping of @bazel_tools does not have a dependency on the //external package, so there is still no circular SkyValue dependency.

Fixes https://github.com/bazelbuild/bazel/issues/15317.

PiperOrigin-RevId: 450665054