[5.2] Remote: Fix a bug that outputs of actions tagged with no-remote are u... (#15453)
* Remote: Fix a bug that outputs of actions tagged with no-remote are u…

…ploaded to remote cache when remote execution is enabled.

Fixes https://github.com/bazelbuild/bazel/issues/14900.

Also fixes an issue that action result from just remotely executed action is not saved to disk cache. The root cause is the action result is inlined in the execution response hence not downloaded through remote cache, hence not saved to disk cache. This results in the second build misses the disk cache, but it can still hit the remote cache and fill the disk cache. The third build can hit disk cache.

Closes #15212.

PiperOrigin-RevId: 441426469

* Fix test

Co-authored-by: Chenchu K <ckolli@google.com>