Add `--starlark_coverage_report`
If the new flag `--starlark_coverage_report` is set to a non-empty
value, the Starlark coverage recorder will be enabled for all `.bzl`
files under the output base executed during the Blaze command. At the
end of the command, the collected coverage data will be emitted to the
file path given as the flag argument in LCOV format.