Add -no_deduplicate for darwin dbg/fastbuild
The -no_deduplicate linker flag saves some time for debug builds at the
cost of binary size. In testing envoy the difference was 50s -> 9s to
link a debug build with this flag enabled. This change enables it as a
separate feature so it can be opted out of by users if needed, and
leaves the default for opt to get this optimization for release builds.

Fixes https://github.com/bazelbuild/bazel/issues/10910