Move information whether the build is using in-memory action file system from `ActionExecutedEvent` to `BuildStartingEvent`.
A build will either use in-memory file system for all actions or none. Move the indicator for this from `ActionExecutedEvent` to `BuildStartingEvent` so that subscibers can note that information when execution starts and use it when receiving events. This way, we can take advantage of this information for other kinds of events.

PiperOrigin-RevId: 453749440
Change-Id: I15dfab0a4af492cbd847dab419f3e13508a2556f