Make the exec group of the test runner configurable
For plain rules, it is possible to give multiple exec groups to a rule
and have some logic inside the rule to pick between an exec group. For
example:

    def _impl(ctx):
        use_large_worker = some == complex && starlark || expression
        ctx.actions.run(
            exec_group = "large" if use_large_worker else "small",
            ...,
        )

For test rules this is only possible for the actions that are run during
the build phase. For actual test execution it is assumed that the exec
group is equal to "test". This means that a construct like the one above
cannot be used to programmatically determine whether a test needs to be
executed on a small or large worker.

This change attempts to remove this limitation by giving
testing.ExecutionInfo a new exec_group argument. It works the same way
as ctx.actions.run(), where it controls the name of the exec group to
use. The default value of this argument is of course "test", for
compatibility with existing code.

Example usage:

    def _impl(ctx):
        use_large_worker = some == complex && starlark || expression
        return [
            DefaultInfo(...),
            testing.ExecutionInfo(
                {},
                exec_group = "large" if use_large_worker else "small",
            ),
        )