Include symlink targets in inputs to Android data binding actions.
The intended semantics of a symlink as an action input is that only the
symlink itself is an input, not the file/directory it points to. However,
due to sandboxing limitations, such an action still executes successfully
locally, though not remotely.

Note that this change is necessary but not sufficient to make these actions
compatible with remote execution; we also need to fix #15678.