Add interface_library to cc_shared_library output group
In general access to the interface library artifact should not be done by the
user and instead the Bazel rules themselves will choose the interface library
file when linking. This is possible if we use C++ rules along the whole
dependency chain.

However, cc_binary(linkshared=1) supported the use case where the actual files
get unwrapped from the rule and it provided access to the interface library via
the `interface_library` output group. This CL adds support for the same thing.

This was only available on Windows for cc_binary and for now will only be
available for Windows on cc_shared_library as well.

RELNOTES:none
PiperOrigin-RevId: 454862408
Change-Id: Iddcffa65be2bfdebac444f30d85b449ad67723e6