Automated rollback of commit 5de967bb27ebdfd31bfa193dc2698cbe46600815.
*** Reason for rollback ***

Temporarily rolling back until #14852 is fixed.

*** Original change description ***

Use the proper main repo mapping where appropriate

Use the main repo mapping (instead of ALWAYS_FALLBACK) to a) process .bzl load labels in the WORKSPACE file, and b) to process any labels passed to repo rules in either the WORKSPACE file or WORKSPACE-loaded macros.

This change only has a noticeable effect for when the workspace name (specified using the `workspace` directive in the WORKSPACE file) is used in labels as `@workspace_name//

***

PiperOrigin-RevId: 454899732
Change-Id: Ief9b90f981d3fdb551c53e56e9ffcfdfb9f61bd4