Make options parsing in BlazeRuntimeWrapper more realistic
Actually select the options for the command and apply the command-specific
invocation policy. Previously everything was considered "build" and the
command passed to the invocation policy was "null", which is not what happens
in production code.

Making this work also required registering commands so that the invocation
policy's command-inheritance logic actually worked. The list of commands isn't
exhaustive, but should do the trick, easy enough to add more as we need them
(I don't think we need _all_ the commands for now at least).

There's still plenty to clean up around options handling in RuntimeWrapper et
al - for example we should be able to pluck options classes from the registered
modules instead of explicitly providing them, but that's for another change.

PiperOrigin-RevId: 455110823
Change-Id: I6223f7508c785e52a160ff41739b704da5354b12