When rewinding actions that produce tree artifacts, invalidate the cached directory creation.
Since the action is to be rewound, it will completely re-execute including `Action#prepare`, which deletes its outputs including tree directories. Invalidate these directories in the `knownDirectories` cache so that they are created again.

This change only affects builds which use rewinding without an in-memory action filesystem.

This will soon be integration tested as part of b/229395817 - in fact, that's how I encountered this bug.

PiperOrigin-RevId: 455250588
Change-Id: I5ca188c95d4528d7fbbfcd108b0b2192f2321d8d