Changed the type param of ActionGraphQueryEnvironment from ConfiguredTargetValue to KeyedConfiguredTargetValue.
ActionGraphQueryEnvironment was reconstructing the ConfiguredTargetKey, given a
ConfiguredTarget. This works fine until the ConfiguredTargetKey is an instance
of ToolchainDependencyConfiguredTargetKey: it can't be reconstructed with
ConfiguredTargetValue since we don't have any information on the execution
platform label.

This CL introduces KeyedConfiguredTargetValue, a wrapper around a
ConfiguredTargetValue that also includes the original ConfiguredTargetKey.
KeyedConfiguredTargetValue is meant for use in aquery only.

A similar solution was implemented for cquery in https://github.com/bazelbuild/bazel/commit/34d98234324da83e93ba0d5ef5702880d5ac7c5c.

PiperOrigin-RevId: 456492858
Change-Id: I1cb5a3193e7283ddd429f4f91b55884ed30c3d9c