Use a single thread pool for transitive ActionLookupValue collection.
With the prior implementation, a thread pool is spun up per top level target when we collect the transitive action lookup values for conflict checking. This is sub-optimal because 1) we pay the setup cost for each pool and 2) there's more contention for resources as there are more threads across pools.

This CL changes the way we collect transitive action lookup values to use 1 single thread pool throughout the whole duration of the build, with the goal to reduce contention.

Code changes:
- Changed TransitiveActionLookupKeysCollector to use FJP directly instead of via a QuiescingExecutor.
- Added IncrementalTransitiveActionLookupKeysCollector, the Skymeld-specific class to collect the transitive ALK that extends TransitiveActionLookupKeysCollector. This instance lives for the entirety of the build, and gets cleaned up afterwards.

PiperOrigin-RevId: 456808137
Change-Id: I63be3cccc9876479072c25634b0266aa80ff4824